read anharmonicity from backend properties
